+++ include/terrain_noise.hh
@@ -11,7 +11,7 @@ public:
         noise_.SetSeed(seed);
     }
 
-    int get_noise(size_t x, size_t y)
+    int get_noise(int x, int y)
     {
         double terrain = sum_octave(16, x, y, 0.4, 0.1, 0, 255);
         return static_cast<int>(terrain);
